// tms-frontend/app/interfaces/auth-registration-data.model.ts
export interface StudentRegistration {
  firstName: string;
  lastName: string;
  email: string;
  mobileNo: string;
  password: string;
  questionId: number;
  answer: number;
}

export interface TutorRegistration {
  firstName: string;
  lastName: string;
  email: string;
  mobileNo: string;
  password: string;
  technology: string;
  about: string;
  questionId: number;
  answer: number;
}